% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatLR.human}
\alias{CellChatLR.human}
\title{Ligand-receptor interactions in CellChat database for human}
\format{
A dataframe includes the ligand-receptor interactions
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
CellChatLR.human
}
\description{
Ligand-receptor interactions in CellChat database for human
}
\keyword{datasets}
