% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatLR.mouse}
\alias{CellChatLR.mouse}
\title{Ligand-receptor interactions in CellChat database for mouse}
\format{
A dataframe includes the ligand-receptor interactions
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
CellChatLR.mouse
}
\description{
Ligand-receptor interactions in CellChat database for mouse
}
\keyword{datasets}
