% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatLR.zebrafish}
\alias{CellChatLR.zebrafish}
\title{Ligand-receptor interactions in CellChat database for zebrafish}
\format{
A dataframe includes the ligand-receptor interactions
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
CellChatLR.zebrafish
}
\description{
Ligand-receptor interactions in CellChat database for zebrafish
}
\keyword{datasets}
