% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{check_Pathway}
\alias{check_Pathway}
\title{find the functionally related pathways in the exSigNet}
\usage{
check_Pathway(Graph, LR.DB, Par)
}
\arguments{
\item{Graph}{exSigNet object with gene expressions and signaling strengths}

\item{LR.DB}{user-select L-R database (here we use CellChatDB)}

\item{Par}{parameters for calculating AcIs, the default values are 2}
}
\description{
find the functionally related pathways in the exSigNet
}
