% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compare_MaxFlow}
\alias{compare_MaxFlow}
\title{calculate the total outflows(inflows) of ligands(targets)}
\usage{
compare_MaxFlow(df)
}
\arguments{
\item{df}{dataframe, output of "get_LRMaxFlow"}
}
\description{
calculate the total outflows(inflows) of ligands(targets)
}
