% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{filterLR_exSigNet}
\alias{filterLR_exSigNet}
\title{find the receptors that only activated by external signals}
\usage{
filterLR_exSigNet(Graph, Exp.Data, DB)
}
\arguments{
\item{Graph}{exSigNet object (gene expressions and signaling strengths are optional)}

\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{DB}{the exFINDER-DB for this species (human, mouse, or zebrafish)}
}
\description{
find the receptors that only activated by external signals
}
