% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fix_OverLap}
\alias{fix_OverLap}
\title{fix the overlap of the external signal-target signaling network (exSigNet) if needed}
\usage{
fix_OverLap(Graph)
}
\arguments{
\item{Graph}{exSigNet object (output of "get_exSigNet")}
}
\description{
fix the overlap of the external signal-target signaling network (exSigNet) if needed
}
