% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_AcI}
\alias{get_AcI}
\title{calculate activation level (AcI) of a given exSigNet}
\usage{
get_AcI(Graph, Par)
}
\arguments{
\item{Graph}{exSigNet object with gene expressions and signaling strengths}

\item{Par}{parameter vector (1x4) for calculating the AcI, the default values are 2}
}
\description{
calculate activation level (AcI) of a given exSigNet
}
