% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basiccomputation.R
\name{get_AveExp}
\alias{get_AveExp}
\title{calculate average expression levels of input genes in every cluster}
\usage{
get_AveExp(Gene, Exp.Data, Meta.Data)
}
\arguments{
\item{Gene}{user-select genes (a list of gene symbols)}

\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{Meta.Data}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}
}
\description{
calculate average expression levels of input genes in every cluster
}
