% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_EdgeWeight}
\alias{get_EdgeWeight}
\title{predict the signaling strengths (edge weights)}
\usage{
get_EdgeWeight(Graph, Kh)
}
\arguments{
\item{Graph}{exSigNet object with gene expressions (output of "get_NodeValue")}

\item{Kh}{parameter for the mass action law (default value is 2)}
}
\description{
predict the signaling strengths (edge weights)
}
