% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_EnrichAnalysis}
\alias{get_EnrichAnalysis}
\title{perform enrichment analysis of a exSigNet}
\usage{
get_EnrichAnalysis(
  Graph,
  OrgDb,
  Number.of.Terms,
  keyType = "ENTREZID",
  ont = "MF",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE
)
}
\arguments{
\item{Graph}{exSigNet object with gene expressions and signaling strengths}

\item{OrgDb}{annotation db}

\item{Number.of.Terms}{output GO terms based on the p-values}

\item{keyType}{keytype of input gene, default is "ENTREZID"}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report, default is 0.05}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported. default is 0.2}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing, default is 10}

\item{maxGSSize}{maximal size of genes annotated for testing, default is 500}

\item{readable}{whether mapping gene ID to gene Name, default is "FALSE"}

\item{pool}{If ont='ALL', whether pool 3 GO sub-ontologies, default is "FALSE"}
}
\description{
perform enrichment analysis of a exSigNet
}
