% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_exSigNet}
\alias{get_exSigNet}
\title{infer the external signal-target signaling network (exSigNet)}
\usage{
get_exSigNet(Graph, Ligands, Exp.Data, Meta.Data, par.cutoff, AG.R, AG.TF)
}
\arguments{
\item{Graph}{ltGRN object (output of "get_ltGRN")}

\item{Ligands}{potential external signals}

\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{Meta.Data}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{par.cutoff}{thresholds to filter the ligands. Negative cutoff value: selecting the ligands below this threshold; zero cutoff value: keep all ligands; otherwise selecting the ligands above this threshold}

\item{AG.R}{list of cell groups that forces exFINDER to filter the receptors based on these cell groups (optional)}

\item{AG.TF}{list of cell groups that forces exFINDER to filter the TFs based on these cell groups (optional)}
}
\description{
uses scRNA-seq data and ltGRN to infer highly expressed TF and R, identify external signals, and infer the exSigNet
}
