% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_exSigNetExData}
\alias{get_exSigNetExData}
\title{infer external signals (and exSigNet) that come from user-defined external cells}
\usage{
get_exSigNetExData(Graph, Exp.ExData, Meta.ExData, cutoff.ExData, AG.ExData)
}
\arguments{
\item{Graph}{exSigNet object (output of "get_exSigNet")}

\item{Exp.ExData}{expression matrix of the external cells (rows: features, columns: cells)}

\item{Meta.ExData}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{cutoff.ExData}{a threshold to filter the ligands (based on the Exp.ExData). Negative cutoff value: selecting the ligands below this threshold; zero cutoff value: keep all ligands; otherwise selecting the ligands above this threshold}

\item{AG.ExData}{list of cell groups (in the external cells) that forces exFINDER to filter the ligands based on these cell groups (optional)}
}
\description{
infer external signals (and exSigNet) that come from user-defined external cells
}
