% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_heatmap}
\alias{get_heatmap}
\title{plot heatmap of the gene expressions in clusters}
\usage{
get_heatmap(
  Gene,
  Exp.Data,
  Meta.Data,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  cellwidth = 15,
  cellheight = 5,
  gaps_col = NULL,
  gaps_row = NULL,
  fontsize_row = 10,
  fontsize_col = 10,
  angle_col = 0,
  legend_breaks = NULL,
  legend_labels = NULL,
  annotation_row = NULL,
  annotation_col = NULL
)
}
\arguments{
\item{Gene}{exSigNet object (gene expressions and signaling strengths are optional)}

\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{Meta.Data}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{cluster_rows}{boolean values determining if rows should be clustered or hclust object, the default value is "FALSE"}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object, the default value is "FALSE"}

\item{cellwidth}{individual cell width in points, the default value is 15}

\item{cellheight}{individual cell height in points, the default value is 5}

\item{gaps_col}{similar to gaps_row, but for columns, the default value is "NULL"}

\item{gaps_row}{vector of row indices that show where to put gaps into heatmap. Used only if the rows are not clustered, the default value is "NULL"}

\item{fontsize_row}{fontsize for rownames, the default value is 10}

\item{fontsize_col}{fontsize for colnames, the default value is 10}

\item{angle_col}{angle of the column labels, right now one can choose only from few predefined options (0, 45, 90, 270 and 315), the default value is 0}

\item{legend_breaks}{vector of breakpoints for the legend}

\item{legend_labels}{vector of labels for the legend_breaks}

\item{annotation_row}{data frame that specifies the annotations shown on left side of the heatmap. Each row defines the features for a specific row. The rows in the data and in the annotation are matched using corresponding row names. Note that color schemes takes into account if variable is continuous or discrete}

\item{annotation_col}{similar to annotation_row, but for columns}
}
\description{
plot heatmap of the gene expressions in clusters
}
