% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_LRMaxFlow}
\alias{get_LRMaxFlow}
\title{calculate the maximal flow between individual ligand and target}
\usage{
get_LRMaxFlow(Graph)
}
\arguments{
\item{Graph}{exSigNet object with gene expressions and signaling strengths}
}
\description{
calculate the maximal flow between individual ligand and target
}
