% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_ltGRN}
\alias{get_ltGRN}
\title{infer ligand-target GRN from user-select target genes based on prior knowledge (exFINDER-DB)}
\usage{
get_ltGRN(Target, DB)
}
\arguments{
\item{Target}{user-select target genes (a list of gene symbols)}

\item{DB}{the exFINDER-DB for this species (human, mouse, or zebrafish)}
}
\description{
infer ligand-target GRN from user-select target genes based on prior knowledge (exFINDER-DB)
}
