% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_NetworkDetail}
\alias{get_NetworkDetail}
\title{show the details of a exSigNet (or ltGRN), number of nodes/edges, etc.}
\usage{
get_NetworkDetail(Graph)
}
\arguments{
\item{Graph}{exSigNet object (output of "get_exSigNet")}
}
\description{
show the details of a exSigNet (or ltGRN), number of nodes/edges, etc.
}
