% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_NodeValue}
\alias{get_NodeValue}
\title{calculate the expression levels of Receptors, TFs, and Targets}
\usage{
get_NodeValue(
  Graph,
  Exp.Data,
  Meta.Data,
  AG.R,
  AG.TF,
  AG.T,
  Exp.ExData,
  Meta.ExData,
  AG.ExData
)
}
\arguments{
\item{Graph}{exSigNet object (output of "get_exSigNet")}

\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{Meta.Data}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{AG.R}{list of cell groups that forces exFINDER to filter the receptors based on these cell groups (optional)}

\item{AG.TF}{list of cell groups that forces exFINDER to filter the TFs based on these cell groups (optional)}

\item{AG.T}{list of cell groups that forces exFINDER to filter the targets based on these cell groups (optional)}

\item{Exp.ExData}{expression matrix of the external cells (rows: features, columns: cells)}

\item{Meta.ExData}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{AG.ExData}{list of cell groups (in the external cells) that forces exFINDER to filter the ligands based on these cell groups (optional)}
}
\description{
option: uses external dataset to calculate the expression levels of the external signals (otherwise they will be set to 1).
}
