% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basiccomputation.R
\name{get_percentile}
\alias{get_percentile}
\title{calculate the precentiel values of the cell groups in a dataset}
\usage{
get_percentile(Exp.Data, Meta.Data, percentile)
}
\arguments{
\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{Meta.Data}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{percentile}{a vector of probabilities (default is c(.5, .75, .90))}
}
\description{
calculate the precentiel values of the cell groups in a dataset
}
