% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_potentialex}
\alias{get_potentialex}
\title{infer the potential external signals based on ltGRN and expression data}
\usage{
get_potentialex(Graph, Exp.Data, Meta.Data, cutoff, AG.L)
}
\arguments{
\item{Graph}{ltGRN object (output of "get_ltGRN")}

\item{Exp.Data}{expression matrix (rows: features, columns: cells)}

\item{Meta.Data}{meta data of the expression matrix, must have a variable "Type" containing all cell labels}

\item{cutoff}{a threshold to filter the ligands. Negative cutoff value: selecting the ligands below this threshold; zero cutoff value: keep all ligands; otherwise selecting the ligands above this threshold}

\item{AG.L}{list of cell groups that forces exFINDER to filter the ligands based on these cell groups (optional)}
}
\description{
uses ltGRN, scRNA-seq data and meta data as inputs to (1) filter out the non-measured genes, (2) calculates average expressions of the ligands in every cluster, (3) infers potential external signals based on their expressions
}
