% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_readyforplot}
\alias{get_readyforplot}
\title{prepare the graph for hierarchy plot}
\usage{
get_readyforplot(Graph)
}
\arguments{
\item{Graph}{exSigNet object (gene expressions and signaling strengths are required)}
}
\description{
prepare the graph for hierarchy plot
}
