% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basiccomputation.R
\name{get_RowMax}
\alias{get_RowMax}
\title{select the highest expression level among all clusters based on the average expression levels (data.frame)}
\usage{
get_RowMax(AveExp)
}
\arguments{
\item{AveExp}{dataframe of averagen expression levels of genes in clusters, output of "get_AveExp"}
}
\description{
select the highest expression level among all clusters based on the average expression levels (data.frame)
}
