% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_SimilarityMatrix}
\alias{get_SimilarityMatrix}
\title{prepare the matrix for k-mean clustering analysis}
\usage{
get_SimilarityMatrix(Graph)
}
\arguments{
\item{Graph}{exSigNet object with gene expressions and signaling strengths}
}
\description{
prepare the matrix for k-mean clustering analysis
}
