% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_subNet}
\alias{get_subNet}
\title{find the sub-exSigNet based on additional TF information}
\usage{
get_subNet(Graph, TF)
}
\arguments{
\item{Graph}{exSigNet object (gene expressions and signaling strengths are optional)}

\item{TF}{user-select TFs}
}
\description{
find the sub-exSigNet based on additional TF information
}
