% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR_layer1_human}
\alias{LR_layer1_human}
\title{Ligand-receptor interactions in exFINDER database for human}
\format{
A dataframe includes the ligand-receptor interactions
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
LR_layer1_human
}
\description{
Ligand-receptor interactions in exFINDER database for human
}
\keyword{datasets}
