% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR_layer1_mouse}
\alias{LR_layer1_mouse}
\title{Ligand-receptor interactions in exFINDER database for mouse}
\format{
A dataframe includes the ligand-receptor interactions
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
LR_layer1_mouse
}
\description{
Ligand-receptor interactions in exFINDER database for mouse
}
\keyword{datasets}
