% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR_layer1_zebrafish}
\alias{LR_layer1_zebrafish}
\title{Ligand-receptor interactions in exFINDER database for zebrafish}
\format{
A dataframe includes the ligand-receptor interactions
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
LR_layer1_zebrafish
}
\description{
Ligand-receptor interactions in exFINDER database for zebrafish
}
\keyword{datasets}
