% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTF_layer2_human}
\alias{RTF_layer2_human}
\title{Receptor-TF signaling in exFINDER database for human}
\format{
A dataframe includes the Receptor-TF signaling
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
RTF_layer2_human
}
\description{
Receptor-TF signaling in exFINDER database for human
}
\keyword{datasets}
