% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTF_layer2_mouse}
\alias{RTF_layer2_mouse}
\title{Receptor-TF signaling in exFINDER database for mouse}
\format{
A dataframe includes the Receptor-TF signaling
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
RTF_layer2_mouse
}
\description{
Receptor-TF signaling in exFINDER database for mouse
}
\keyword{datasets}
