% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFT_layer3_human}
\alias{TFT_layer3_human}
\title{TF-Target regulation in exFINDER database for human}
\format{
A dataframe includes the TF-Target regulations
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
TFT_layer3_human
}
\description{
TF-Target regulation in exFINDER database for human
}
\keyword{datasets}
