% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFT_layer3_zebrafish}
\alias{TFT_layer3_zebrafish}
\title{TF-Target regulation in exFINDER database for zebrafish}
\format{
A dataframe includes the TF-Target regulations
}
\source{
\url{https://github.com/ChanghanGitHub/exFINDER/}
}
\usage{
TFT_layer3_zebrafish
}
\description{
TF-Target regulation in exFINDER database for zebrafish
}
\keyword{datasets}
